requirements.txt
couchdb-requests==0.0.1
docopt==0.6.2
requests==2.7.0
